Fix connecting to an already running target in VSCode
When a GDB client first connects, it assumes the current Thread ID is
0xDEAD. Return this dummy Thread ID until GDB retrieves the real
current Thread ID. Return to the uninitialized state whenever
RTOS_GetSymbols is called or when RTOS_GetThreadDisplay is called with
the dummy 0xDEAD thread ID to account for new uninitialized clients
connecting.
Without this, VS Code's GDB client either fails to initialize or fails
to single step when connecting to a target with the kernel already
initialized.

Signed-off-by: Josh DeWitt <josh.dewitt@garmin.com>